Cleaned controller and added a game logic modell
